add 8 chart

+++ src/AntDesign.Charts/Components/Configs/IValueAxis.cs
@@ -12,8 +12,8 @@ namespace AntDesign.Charts
         ///  tick相关配置 
         /// </summary>
         public bool? nice { get; set; }
-        public int? min { get; set; }
-        public int? max { get; set; }
+        public double? min { get; set; }
+        public double? max { get; set; }
         public int? minLimit { get; set; }
         public int? maxLimit { get; set; }
         public int? tickCount { get; set; }
@@ -34,8 +34,8 @@ namespace AntDesign.Charts
     {
         public string type { get; set; }
         public bool? nice { get; set; }
-        public int? min { get; set; }
-        public int? max { get; set; }
+        public double? min { get; set; }
+        public double? max { get; set; }
         public int? minLimit { get; set; }
         public int? maxLimit { get; set; }
         public int? tickCount { get; set; }
